refactor: Separate curve iterator from MultiCurve
The main purpose is to start generalizing the scribbler so that
it doesn't limit itself to lines. As a result, we had to tear
the MultiCurve interface apart, so that the Curve iterator is
a separate entity and the MultiCurve is essentially a collection
of curves and a bounding box.